requirements_updated.txt
altair==5.5.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
decorator==4.4.2
Dehaze==0.0.4
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
humanfriendly==10.0
idna==3.10
image-dehazer==0.0.9
imageio==2.36.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
narwhals==1.18.4
networkx==3.4.2
numba==0.60.0
numpy==1.25.0
onnxruntime==1.20.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
protobuf==5.29.1
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.18.0
PyMatting==1.1.13
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rembg==2.0.60
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.14.1
six==1.17.0
smmap==5.0.1
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
tifffile==2024.12.12
toml==0.10.2
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
